# src/main/resources/specs/collector-manager-api-spec.yaml
openapi: 3.0.3
info:
  title: Recolnat
  description: |-
    Le réseau national des collections naturalistes 
    est une infrastructure de recherche française dont l'objectif est 
    de produire un corpus de données basé sur les collections d'histoire naturelle de France, 
    permettant de valoriser les recherches au service de l'étude de la géo- et de la biodiversité actuelle et passée.
    Liens utiles:
    - [RECOLNAT](https://fr.wikipedia.org/wiki/RECOLNAT)
  version: "1.0"
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@mnhn.fr
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: https://petstore3.swagger.io/api/v3
...
security:
  - collection-oidc: [ ] # <--- Use the same name as specified in securitySchemes*, here without scope
paths:
...
  /v1/import/identification/template/generate:
    post:
      operationId: generateIdentificationTemplate
      tags:
        - ImportIdentification
      summary: Génère le modèle de fichier d'import de déterminations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
...
  /v1/import/identification/template:
    get:
      operationId: getIdentificationTemplateStructure
      tags:
        - ImportIdentification
      summary: Récupère le modèle de fichier d'import de déterminations
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateStructure'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
...
  /v1/import/identification/check:
    post:
      operationId: checkIdentifications
      tags:
        - ImportIdentification
      summary: Vérification des données de détermination
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  description: Fichier à uploader
                  type: string
                  format: binary
                institutionId:
                  description: Identifiant de l'institution sur laquelle on veut faire l'import
                  type: string
                  format: uuid
                  example: 386ff0b0-a6cc-11ec-b909-0242ac120002
              required:
                - file
                - institutionId
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportCheckDataResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
...
  /v1/articles:
    get:
      operationId: getArticles
      parameters:
        - name: page
          required: true
          in: query
          schema:
            type: integer
            format: int32
            minimum: 0
            description: number of page to retrieve
          example:
            id: 1
        - name: size
          required: true
          in: query
          schema:
            type: integer
            format: int32
            description: number of elements per page to retrieve
          example:
            id: 10
        - name: q
          required: false
          in: query
          description: String to look for in article title
          schema:
            type: string
        - name: state
          required: false
          in: query
          description: State of the article
          schema:
            $ref: '#/components/schemas/ArticleState'
        - name: columnSort
          required: false
          in: query
          description: Champ de tri
          schema:
            type: string
        - name: typeSort
          in: query
          description: Ordre de tri
          schema:
            type: string
            enum:
              - ASC
              - DESC
      tags:
        - Article
      summary: Retrieve articles
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleDashboardResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    post:
      operationId: addArticle
      tags:
        - Article
      summary: add Article
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - content
                - title
                - author
                - creationDate
                - state
              properties:
                title:
                  type: string
                content:
                  type: string
                author:
                  type: string
                creationDate:
                  type: string
                  description: date creation article , format YYYY-MM-JJ
                  example: 2023-12-18
                media:
                  type: string
                  format: binary
                state:
                  $ref: "#/components/schemas/ArticleState"
      responses:
        '201':
          description: Successful creation
          headers:
            articleId:
              description: article identifier
              required: true
              schema:
                type: string
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/articles/{id}:
    get:
      operationId: getArticleById
      tags:
        - Article
      summary: get Article By Id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: article identifier
            example:
              id: 386ff0b0-a6cc-11ec-b909-0242ac120002

      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    put:
      operationId: UpdateArticleById
      tags:
        - Article
      summary: Update Article by Id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: article identifier
            example:
              id: 386ff0b0-a6cc-11ec-b909-0242ac120002
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Article'
      responses:
        '200':
          description: Successful creation
          headers:
            articleId:
              description: article identifier
              required: true
              schema:
                type: string
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
    delete:
      operationId: deleteArticle
      tags:
        - Article
      summary: Delete an article
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: article identifier
            example:
              id: 386ff0b0-a6cc-11ec-b909-0242ac120002
      responses:
        '200':
          description: Successful creation
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/articles/{id}:
    get:
      operationId: getPublicArticleById
      tags:
        - visibilite public
      summary: get Article By Id (for public acces)
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
            format: uuid
            description: article identifier
            example:
              id: 386ff0b0-a6cc-11ec-b909-0242ac120002

      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
  /v1/public/articles:
    get:
      operationId: getPublicArticles
      tags:
        - visibilite public
      summary: Retrieve all articles by pages  (for public acces)
      parameters:
        - in: query
          required: true
          name: page
          schema:
            type: integer
            format: int32
          description: page to access
        - in: query
          required: true
          name: size
          schema:
            type: integer
            format: int32
          description: size of elements per page
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResultPageResponse'
        '401':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '409':
          description: Applicative Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
        '500':
          description: error server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnexpectedError'
...
components:
  securitySchemes:
    collection-oidc: # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: http://sample.com/realms/recolnat-oauth2/.well-known/openid-configuration
  schemas:
...
    ArticleState:
      type: string
      enum:
        - PUBLISHED
        - DRAFT
        -
...
    ArticleDashboardResponse:
      description: Article Response
      type: object
      nullable: false
      properties:
        totalPages:
          type: string
          format: int
        numberOfElements:
          type: string
          format: long
        data:
          items:
            $ref: "#/components/schemas/ArticleDashboard"
...
    ArticleDashboard:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Id de l'article
          example: 50f4978a-da62-4fde-8f38-5003bd43ff64
        title:
          type: string
          format: string
          description: Titre de l'article
          example: Un titre
        author:
          type: string
          format: string
          description: Auteur de l'article
          example: John Doe
        creationDate:
          type: string
          format: date
          description: Date de création de l'article
        state:
          $ref: '#/components/schemas/ArticleState'
...
    Article:
      type: object
      required:
        - content
        - author
        - title
        - creationDate
      properties:
        title:
          type: string
          description: titre de l article
          example: Mon titre
        content:
          type: string
          description: description contenu de l article
          example: Mon contenu
        author:
          type: string
          description: auteur de l article
          example: Mon nom de plume
        creationDate:
          type: string
          description: date creation article , format YYYY-MM-JJ
          example: 2023-12-18
        media:
          type: string
          format: binary
          description: dle binaire de l article
          example: fichier svg
        state:
          $ref: "#/components/schemas/ArticleState"
    ArticleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        author:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        creationDate:
          type: string
          format: date
        urlMedia:
          type: string
        media:
          type: string
          format: binary
        incrementId:
          type: integer
    ArticleResultPageResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        articles:
          type: array
          items:
            $ref: '#/components/schemas/ArticleResponse'
...
    UnexpectedError:
      $ref: '#/components/schemas/Error'
    Forbidden:
      $ref: '#/components/schemas/Error'
    Unauthorized:
      $ref: '#/components/schemas/Error'
    Conflict:
      $ref: '#/components/schemas/Error'
    Error:
      type: object
      description: A common error response
      properties:
        errorCode:
          type: integer
          description: Error Code
        errorMessage:
          type: string
          description: Message
        errorLevel:
          type: string
          description: Level of error
          default: info
          enum:
            - info
            - warning
            - error
            - critical
        errorType:
          type: string
          description: Type of error
          enum:
            - functional
            - technical
        documentationUrl:
          type: string
          description: Url of the documenation
        tips:
          type: string
          description: Tips/help for adressing the error
    ImportCheckDataResponse:
      type: object
      properties:
        format:
          type: boolean
        size:
          type: boolean
        structureErrors:
          type: array
          items:
            $ref: '#/components/schemas/ImportStructureError'
...
    ImportStructureError:
      type: object
      properties:
        code:
          type: string
        columns:
          type: array
          items:
            type: string
...
    TemplateStructure:
      type: object
      properties:
        name:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/TemplateField'
    TemplateField:
      type: object
      properties:
        key:
          type: string
        required:
          type: boolean